move all py files out of pipeline dir
