+++ tests/test_document_file_checksums.py
@@ -1,7 +1,7 @@
 import os
 import filecmp
 from tests.test_sarscov2_consensus_acceptance import FileTestCase
-from pipeline.document_file_checksums import generate_checksums_file
+from qc.document_file_checksums import generate_checksums_file
 
 
 class DocumentFileChecksumsTest(FileTestCase):
